doc-bot/requirements.txt
slack-bolt>=1.18.0
anthropic>=0.25.0
google-auth>=2.29.0
google-api-python-client>=2.128.0
